pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infra-gen"
version = "0.1.0"
description = "Generate production-ready multi-environment infrastructure-as-code from YAML manifests"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
infra-gen = "infra_gen.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

# --- Ruff ---
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "F",      # pyflakes
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["N802", "N803", "B011", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- Mypy ---
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true
